exception

+++ sts/Swager_Demo/src/main/java/com/swager/service/BookService.java
@@ -1,11 +1,15 @@
 package com.swager.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.RequestEntity;
 import org.springframework.stereotype.Service;
 
 import com.swager.dto.BookRequestDTO;
+import com.swager.dto.ResponseDTO;
+import com.swager.exception.BookNotFoundException;
 import com.swager.modal.Book;
 import com.swager.repo.BookRepo;
 
@@ -18,10 +22,10 @@ public class BookService
 	
 		// CRUD operations.,
 		// insert the records,
-		public Book insertBook(BookRequestDTO bookDto)
-		{	
-			Book bk = Book.name(0,bookDto.getBookName(),bookDto.getBookAuthor(),bookDto.getBookPrice());
-			return bookRepo.save(bk);
+		public Book insertBook(BookRequestDTO bookRequestDTO)
+		{
+			Book book = new Book(bookRequestDTO);
+			return bookRepo.save(book);
 		}
 		
 		// insert the List.,
@@ -32,15 +36,31 @@ public class BookService
 		}
 		
 		// get the record,
-		public Book getBook(int bookID)
+		public Optional<Book> getBook(int bookID)
 		{
-			return bookRepo.findById(bookID).get();
+			Optional<Book> book = bookRepo.findById(bookID);
+				if(book.isPresent())
+						{
+							return book;
+						}
+				else
+				{
+					throw new BookNotFoundException("No book was found with the give book ID"+bookID);
+				}
 		}
 		
 		//get record by name., custom query(findByfieldName)
 		public Book getBookByName(String bookName)
 		{
-			return bookRepo.findBybookName(bookName);
+			Book book = bookRepo.findBybookName(bookName);
+					if(bookRepo.findBybookName(bookName)==null)
+					{
+						throw new BookNotFoundException("Provide the proper book name idiot,");
+					}
+					else
+					{
+						return book;
+					}
 		}
 		
 		// get all the records,
@@ -53,18 +73,33 @@ public class BookService
 		public Book updateBook(BookRequestDTO book, int bookID)
 		{
 			Book bk = bookRepo.findById(bookID).get();
-				bk.setBookName(book.getBookName());
-				bk.setBookAuthor(book.getBookAuthor());
-				bk.setBookPrice(book.getBookPrice());
-
-				return bookRepo.save(bk);
+				if(bookRepo.findById(bookID).isPresent())
+				{
+					bk.setBookName(book.getBookName());
+					bk.setBookAuthor(book.getBookAuthor());
+					bk.setBookPrice(book.getBookPrice());
+					
+					return bookRepo.save(bk);
+				}
+				else
+				{
+					throw new BookNotFoundException("No book to update");
+				}
 		}
 		
 		// delete the record
 		public String deleteBook(int bookID)
 		{
-			bookRepo.deleteById(bookID);
-			return "";
+			Optional<Book> book = bookRepo.findById(bookID);
+					if(book.isPresent())
+					{
+						bookRepo.deleteById(bookID);
+						return "";
+					}
+					else
+					{
+						throw new BookNotFoundException("No book with the id to update");
+					}
 		}
 		
 }
